pyproject.toml
[tool.poetry]
name = "stick-me"
version = "0.1.0"
description = "A Telegram bot which helps to create a sticker set with your beloved stickers"
authors = ["Evgenii Korobkov <eug.korobkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

httpx = "==0.27.2"
python-telegram-bot = "==21.6.0"

[tool.poetry.group.dev.dependencies]
mypy = "==1.11.2"
pylint = "==3.3.1"
ruff = "==0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 99
indent-width = 4

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pylint]
max-args = 10
min-public-methods = 0

[tool.mypy]
# Specify the directories to include for type checking.
exclude = ["migrations"]

[tool.pylint.'DESIGN']
max-positional-arguments = 10
